feat: simplify sighash all contract
Notice this commit also deletes all the redundant parts, the reason
for this is that the original bitcoin lock contract is more of an
experiment, upon further development, we decided to go with a minimal
sighash all contract. Hence we are removing all the unnecessary parts
and only keep critical system scripts here.
