--- c/protocol.fbs
@@ -1,243 +1,0 @@
-namespace Ckb.Protocol;
-
-union SyncPayload {
-    GetHeaders,
-    Headers,
-    GetBlocks,
-    Block,
-    SetFilter,
-    AddFilter,
-    ClearFilter,
-    FilteredBlock,
-}
-
-table SyncMessage {
-    payload:        SyncPayload;
-}
-
-table Bytes {
-    seq: [ubyte];
-}
-
-table GetHeaders {
-    version:                uint32;
-    block_locator_hashes:   [H256];
-    hash_stop:              H256;
-}
-
-table GetBlocks {
-    block_hashes:           [H256];
-}
-
-table Headers {
-    headers:                [Header];
-}
-
-table Header {
-    version:                uint32;
-    parent_hash:            H256;
-    timestamp:              uint64;
-    number:                 uint64;
-    transactions_root:      H256;
-    witnesses_root:         H256;
-    proposals_hash:         H256;
-    difficulty:             Bytes;
-    nonce:                  uint64;
-    proof:                  Bytes;
-    uncles_hash:            H256;
-    uncles_count:           uint32;
-    epoch:                  uint64;
-}
-
-table Block {
-    header:                 Header;
-    uncles:                 [UncleBlock];
-    transactions:           [Transaction];
-    proposals:              [ProposalShortId];
-}
-
-table UncleBlock {
-    header:                 Header;
-    proposals:              [ProposalShortId];
-}
-
-table Transaction {
-    version:        uint32;
-    deps:           [OutPoint];
-    inputs:         [CellInput];
-    outputs:        [CellOutput];
-    witnesses:      [Witness];
-}
-
-table Witness {
-    data:           [Bytes];
-}
-
-table OutPoint {
-    block_hash:     H256;
-    tx_hash:        H256;
-    index:          uint32;
-}
-
-table CellInput {
-    block_hash:     H256;
-    tx_hash:        H256;
-    index:          uint32;
-    since:          uint64;
-    args:           [Bytes];
-}
-
-table CellOutput {
-    capacity:       uint64;
-    data:           Bytes;
-    lock:           Script;
-    type:           Script;
-}
-
-table Script {
-    args:           [Bytes];
-    code_hash:      H256;
-}
-
-root_type SyncMessage;
-
-union RelayPayload {
-    CompactBlock,
-    RelayTransaction,
-    RelayTransactionHash,
-    GetRelayTransaction,
-    GetBlockTransactions,
-    BlockTransactions,
-    GetBlockProposal,
-    BlockProposal,
-}
-
-table RelayMessage {
-    payload:        RelayPayload;
-}
-
-table CompactBlock {
-    header:                     Header;
-    nonce:                      uint64;
-    short_ids:                  [Bytes];
-    prefilled_transactions:     [IndexTransaction];
-    uncles:                     [UncleBlock];
-    proposals:                  [ProposalShortId];
-}
-
-table IndexTransaction {
-    index:                      uint32;
-    transaction:                Transaction;
-}
-
-table RelayTransactionHash {
-    tx_hash: H256;
-}
-
-table GetRelayTransaction {
-    tx_hash: H256;
-}
-
-table RelayTransaction {
-    cycles:                     uint64;
-    transaction:                Transaction;
-}
-
-table GetBlockTransactions {
-    block_hash:                H256;
-    indexes:                   [uint32];
-}
-
-table BlockTransactions {
-    block_hash:                H256;
-    transactions:              [Transaction];
-}
-
-table GetBlockProposal {
-    block_number:              uint64;
-    proposals:     [ProposalShortId];
-}
-
-table BlockProposal {
-    transactions:              [Transaction];
-}
-
-struct ProposalShortId {
-    u0: uint8;
-    u1: uint8;
-    u2: uint8;
-    u3: uint8;
-    u4: uint8;
-    u5: uint8;
-    u6: uint8;
-    u7: uint8;
-    u8: uint8;
-    u9: uint8;
-}
-
-struct H256 {
-    u0: uint8;
-    u1: uint8;
-    u2: uint8;
-    u3: uint8;
-    u4: uint8;
-    u5: uint8;
-    u6: uint8;
-    u7: uint8;
-    u8: uint8;
-    u9: uint8;
-    u10: uint8;
-    u11: uint8;
-    u12: uint8;
-    u13: uint8;
-    u14: uint8;
-    u15: uint8;
-    u16: uint8;
-    u17: uint8;
-    u18: uint8;
-    u19: uint8;
-    u20: uint8;
-    u21: uint8;
-    u22: uint8;
-    u23: uint8;
-    u24: uint8;
-    u25: uint8;
-    u26: uint8;
-    u27: uint8;
-    u28: uint8;
-    u29: uint8;
-    u30: uint8;
-    u31: uint8;
-}
-
-table SetFilter {
-    filter: [uint8];
-    num_hashes: uint8;
-    hash_seed: uint32;
-}
-
-table AddFilter {
-    filter: [uint8];
-}
-
-table ClearFilter {
-}
-
-table FilteredBlock {
-    header: Header;
-    transactions: [Transaction];
-    proof: MerkleProof;
-}
-
-table MerkleProof {
-    indices: [uint32];
-    lemmas: [H256];
-}
-
-table TimeMessage {
-    payload:        Time;
-}
-
-table Time {
-    timestamp: uint64;
-}
