--- c/secp256k1_blake160_lock.c
@@ -1,6 +1,0 @@
-#include "secp256k1_blake160.h"
-
-int main(int argc, char* argv[])
-{
-  return verify_bitcoin_sighash(argc, argv);
-}
